image-matcher/requirements.txt
streamlit>=1.31.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
torch>=2.1.0
Pillow>=10.0.0
numpy>=1.24.0